Changes after OSS review 1st pass
