+++ samza-api/src/main/java/org/apache/samza/storage/BlobStoreAdminFactory.java
@@ -4,6 +4,15 @@ import org.apache.samza.config.Config;
 import org.apache.samza.job.model.JobModel;
 
 
+/**
+ * Factory to create instance of {@link StateBackendAdmin}s that needs to be implemented for every
+ * state backend
+ */
 public interface BlobStoreAdminFactory {
+  /**
+   * Returns an instance of {@link StateBackendAdmin}
+   * @param config job configuration
+   * @param jobModel Job Model
+   */
   StateBackendAdmin getStateBackendAdmin(Config config, JobModel jobModel);
 }
