+++ samza-core/src/main/java/org/apache/samza/standalone/PassthroughJobCoordinator.java
@@ -90,18 +90,6 @@ public class PassthroughJobCoordinator implements JobCoordinator {
       // TODO metrics registry has been null here for a while; is it safe?
       MetadataResourceUtil metadataResourceUtil = new MetadataResourceUtil(jobModel, null, config);
       metadataResourceUtil.createResources();
-
-      // create all the resources required for state backend factories
-      new StorageConfig(config).getStateBackendBackupFactories().forEach(stateStorageBackendBackupFactory -> {
-        StateBackendFactory stateBackendFactory =
-            ReflectionUtil.getObj(stateStorageBackendBackupFactory, StateBackendFactory.class);
-        StateBackendAdmin stateBackendAdmin = stateBackendFactory.getStateBackendAdmin(getJobModel(), config);
-        // Create resources required for state backend admin
-        stateBackendAdmin.createResources();
-        // Validate resources required for state backend admin
-        stateBackendAdmin.validateResources();
-      });
-
     } catch (Exception e) {
       LOGGER.error("Exception while trying to getJobModel.", e);
       if (coordinatorListener != null) {
